+++ src/pages/AccountSummaryDemo.vue
@@ -1,0 +1,7 @@
+<template>
+  <q-page class="q-pa-md q-mt-xl"> </q-page>
+</template>
+
+<script setup></script>
+
+<style scoped></style>